pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.5",
    "django-ckeditor>=6.7.3",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "drf-spectacular[sidecar]>=0.28.0",
    "jinja2>=3.1.6",
    "markdown>=3.8.2",
    "orjson>=3.11.2",
    "pillow>=11.3.0",
    "redis>=6.4.0",
]
